:truck: Oppdatert struktur for sanity-moduler
