+++ aksel.nav.no/website/components/index.ts
@@ -1,20 +1,5 @@
 export * from "./assets";
-export { default as Accordion } from "./sanity-modules/Accordion";
-export { default as Alert } from "./sanity-modules/Alert";
-export { default as Bilde } from "./sanity-modules/Bilde";
-export { default as DoDont } from "./sanity-modules/DoDont";
-export { default as IntroKomponent } from "./sanity-modules/IntroSeksjon";
-export { default as SideModul } from "./sanity-modules/SideModul";
-export { default as Tabell } from "./sanity-modules/Tabell";
-export { default as TastaturModul } from "./sanity-modules/TastaturTabell";
-export { default as Tips } from "./sanity-modules/Tips";
-export { default as TokenTable } from "./sanity-modules/TokenTable";
 export * from "./sanity-modules/cards";
-export * from "./sanity-modules/code";
-export * from "./sanity-modules/code-examples";
-export { default as PropTable } from "./sanity-modules/props/PropTabell";
-export { default as PropsSeksjon } from "./sanity-modules/props/PropsSeksjon";
-export { default as Video } from "./sanity-modules/video/Video";
 export * from "./website-modules/BreadCrumbs";
 export { default as Feedback } from "./website-modules/feedback";
 export * from "./website-modules/toc/TOC";
