--- aksel.nav.no/website/components/sanity-modules/code-examples/index.tsx
@@ -1,1 +1,0 @@
-export { default as CodeExamples } from "./Examples";
