+++ aksel.nav.no/website/components/sanity-modules/code-snippet/SnippetLazy.tsx
@@ -1,0 +1,7 @@
+import dynamic from "next/dynamic";
+
+const SnippetLazy = dynamic(() => import("./Snippet"), {
+  ssr: false,
+});
+
+export default SnippetLazy;
