--- aksel.nav.no/website/components/sanity-modules/code/index.tsx
@@ -1,5 +1,0 @@
-import dynamic from "next/dynamic";
-
-export const Snippet = dynamic(() => import("./Snippet"), {
-  ssr: false,
-});
