+++ aksel.nav.no/website/components/sanity-modules/frontpage-blocks/FrontpageBlocks.tsx
@@ -1,5 +1,5 @@
 import ErrorBoundary from "@/error-boundary";
-import Latest, { LatestT } from "./Latest";
+import Latest, { LatestT } from "./latest-articles/Latest";
 export type BlocksT = LatestT;
 
 type FrontpageBlockProps = {
