+++ aksel.nav.no/website/components/sanity-modules/token-view/TokenView.tsx
@@ -1,12 +1,12 @@
 import ErrorBoundary from "@/error-boundary";
-import { BreakpointsView } from "./categories/Breakpoints";
-import { FontView } from "./categories/Font";
-import { GlobalView } from "./categories/Global";
-import { SemanticView } from "./categories/Semantic";
-import { ShadowView } from "./categories/Shadow";
-import { ShapesView } from "./categories/Shapes";
-import { SpacingView } from "./categories/Spacing";
-import { ZindexView } from "./categories/Zindex";
+import { BreakpointsView } from "./parts/categories/Breakpoints";
+import { FontView } from "./parts/categories/Font";
+import { GlobalView } from "./parts/categories/Global";
+import { SemanticView } from "./parts/categories/Semantic";
+import { ShadowView } from "./parts/categories/Shadow";
+import { ShapesView } from "./parts/categories/Shapes";
+import { SpacingView } from "./parts/categories/Spacing";
+import { ZindexView } from "./parts/categories/Zindex";
 
 type TokenViewProps = {
   token: {
