+++ aksel.nav.no/website/pages/grunnleggende/index.tsx
@@ -12,7 +12,7 @@ import cl from "clsx";
 import { WithSidebar } from "components/layout/WithSidebar";
 import Footer from "components/layout/footer/Footer";
 import { Header } from "components/layout/header/Header";
-import ComponentOverview from "components/sanity-modules/ComponentOverview";
+import ComponentOverview from "components/sanity-modules/component-overview/ComponentOverview";
 import { SEO } from "components/website-modules/seo/SEO";
 import { GetStaticProps } from "next/types";
 import { Suspense, lazy } from "react";
