+++ aksel.nav.no/website/pages/komponenter/[...slug].tsx
@@ -22,8 +22,8 @@ import { BodyShort, Detail, Heading } from "@navikt/ds-react";
 import Footer from "components/layout/footer/Footer";
 import { Header } from "components/layout/header/Header";
 import { WithSidebar } from "components/layout/WithSidebar";
-import ComponentOverview from "components/sanity-modules/ComponentOverview";
-import IntroSeksjon from "components/sanity-modules/IntroSeksjon";
+import ComponentOverview from "components/sanity-modules/component-overview/ComponentOverview";
+import IntroSeksjon from "components/sanity-modules/intro-seksjon/IntroSeksjon";
 import { SEO } from "components/website-modules/seo/SEO";
 import { StatusTag } from "components/website-modules/StatusTag";
 import { SuggestionBlock } from "components/website-modules/suggestionblock";
