+++ aksel.nav.no/website/sanity/schema/objects/index.ts
@@ -18,4 +18,3 @@ export * from "./shared/tabell";
 export * from "./shared/tips";
 export * from "./shared/unik-sidemodul";
 export * from "./shared/video";
-export * from "./standalone/uu-feedback";
